Final Front End with Email Entry
Changes applied to the layout style. Container is our main div within the body of the HTML. Within Container we have a header, content, and footer. Header consists of the title, content consists of the chat box and the repID window, and the footer consists of the href links to resource pages. Everything is using flex box. Added email entry for chat transcript.